fix: separate delagate, and use parent indices.
