+++ gudpy/gui/widgets/tables/components_table.py
@@ -9,6 +9,7 @@ from core.composition import Component
 from core.element import Element
 from core.isotopes import Sears91
 from gui.widgets.tables.composition_table import CompositionDelegate
+from gui.widgets.tables.composition_delegate import CompositionDelegate
 
 
 class ComponentsModel(QAbstractItemModel):
